+++ pyriemann/utils/geodesic.py
@@ -1,7 +1,8 @@
 """Geodesics for SPD/HPD matrices."""
 import numpy as np
+from scipy.linalg import eigvalsh
 
-from .base import ctranspose, sqrtm, invsqrtm, powm, logm, expm
+from .base import _recursive, ctranspose, sqrtm, invsqrtm, powm, logm, expm
 from .utils import check_function
 
 
@@ -91,8 +92,7 @@ def geodesic_logchol(A, B, alpha=0.5):
     r"""Log-Cholesky geodesic between SPD/HPD matrices.
 
     The matrix at position :math:`\alpha` on the log-Cholesky geodesic
-    between two SPD/HPD matrices :math:`\mathbf{A}` and :math:`\mathbf{B}` is
-    given in [1]_.
+    between two SPD/HPD matrices :math:`\mathbf{A}` and :math:`\mathbf{B}` is:
 
     Parameters
     ----------
@@ -168,6 +168,14 @@ def geodesic_logeuclid(A, B, alpha=0.5):
     See Also
     --------
     geodesic
+
+    References
+    ----------
+    .. [1] `Geometric means in a novel vector space structure on symmetric
+        positive-definite matrices
+        <https://epubs.siam.org/doi/abs/10.1137/050637996>`_
+        V. Arsigny, P. Fillard, X. Pennec, N. Ayache.
+        SIAM J Matrix Anal Appl, 2007, 29 (1), pp. 328-347
     """
     return expm((1 - alpha) * logm(A) + alpha * logm(B))
 
@@ -203,9 +211,79 @@ def geodesic_riemann(A, B, alpha=0.5):
     See Also
     --------
     geodesic
+
+    References
+    ----------
+    .. [1] `Riemannian geometry and matrix geometric means
+        <https://www.sciencedirect.com/science/article/pii/S0024379505004350>`_
+        R. Bhatia and J. Holbrook.
+        Linear Algebra and its Applications, 2006
     """
     sA, isA = sqrtm(A), invsqrtm(A)
     C = sA @ powm(isA @ B @ isA, alpha) @ sA
+    return C
+
+
+def geodesic_thompson(A, B, alpha=0.5):
+    r"""Thompson geodesic between SPD/HPD matrices.
+
+    The matrix at position :math:`\alpha` on a possible Thompson geodesic
+    between two SPD/HPD matrices :math:`\mathbf{A}` and :math:`\mathbf{B}` is
+    given in [1]_.
+
+    :math:`\mathbf{C}` is equal to :math:`\mathbf{A}` if :math:`\alpha` = 0,
+    and :math:`\mathbf{B}` if :math:`\alpha` = 1.
+
+    Parameters
+    ----------
+    A : ndarray, shape (..., n, n)
+        First SPD/HPD matrices.
+    B : ndarray, shape (..., n, n)
+        Second SPD/HPD matrices.
+    alpha : float, default=0.5
+        Position on the geodesic.
+
+    Returns
+    -------
+    C : ndarray, shape (..., n, n)
+        SPD/HPD matrices on the Thompson geodesic.
+
+    Notes
+    -----
+    ..versionadded:: 0.10
+
+    See Also
+    --------
+    geodesic
+
+    References
+    ----------
+    .. [1] `Differential geometry with extreme eigenvalues in the positive
+        semidefinite cone
+        <https://arxiv.org/pdf/2304.07347>`_
+        C. Mostajeran, N. Da Costa, G. Van Goffrier and R. Sepulchre.
+        SIAM Journal on Matrix Analysis and Applications, 2024
+    """
+    A_ndim = A.ndim
+    while A.ndim < 3:
+        A, B = A[np.newaxis, ...], B[np.newaxis, ...]
+    E = _recursive(eigvalsh, B, A)
+    Emin, Emax = E.min(axis=-1), E.max(axis=-1)
+
+    C = np.zeros_like(A)
+
+    mask = (Emin == Emax)
+    C[mask] = (Emin[mask][..., np.newaxis, np.newaxis] ** alpha) * A[mask]
+
+    Emin_a, Emax_a = Emin ** alpha, Emax ** alpha
+    b = (Emax_a - Emin_a)[~mask][..., np.newaxis, np.newaxis]
+    a = (Emax * Emin_a - Emin * Emax_a)[~mask][..., np.newaxis, np.newaxis]
+    c = b * B[~mask] + a * A[~mask]
+    C[~mask] = c / (Emax - Emin)[~mask][..., np.newaxis, np.newaxis]
+
+    if A_ndim < 3:
+        C = C[0]
+
     return C
 
 
@@ -249,14 +327,14 @@ def geodesic_wasserstein(A, B, alpha=0.5):
     ----------
     .. [1] `Wasserstein Riemannian geometry of Gaussian densities
         <https://link.springer.com/article/10.1007/s41884-018-0014-4>`_
-        L. Malagò, L. Montrucchio, G. Pistone. Information Geometry, 2018, 1,
-        pp. 137–179.
+        L. Malagò, L. Montrucchio, G. Pistone.
+        Information Geometry, 2018, 1, pp. 137–179.
     """
     A12 = sqrtm(A)
     A12inv = invsqrtm(A)
     AB12 = A12 @ sqrtm(A12 @ B @ A12) @ A12inv
     return (1-alpha)**2 * A + alpha**2 * B + \
-        alpha*(1-alpha) * (AB12 + AB12.conj().swapaxes(-1, -2))
+        alpha*(1-alpha) * (AB12 + ctranspose(AB12))
 
 
 ###############################################################################
@@ -268,15 +346,19 @@ geodesic_functions = {
     "logchol": geodesic_logchol,
     "logeuclid": geodesic_logeuclid,
     "riemann": geodesic_riemann,
+    "thompson": geodesic_thompson,
     "wasserstein": geodesic_wasserstein,
 }
 
 
 def geodesic(A, B, alpha, metric="riemann"):
-    """Geodesic between matrices according to a metric.
+    r"""Geodesic between matrices according to a metric.
 
     Return the matrix at the position alpha on the geodesic between matrices
     A and B according to a metric.
+
+    :math:`\mathbf{C}` is equal to :math:`\mathbf{A}` if :math:`\alpha` = 0,
+    and :math:`\mathbf{B}` if :math:`\alpha` = 1.
 
     Parameters
     ----------
@@ -288,7 +370,8 @@ def geodesic(A, B, alpha, metric="riemann"):
         Position on the geodesic.
     metric : string | callable, default="riemann"
         Metric used for geodesic, can be:
-        "chol", "euclid", "logchol", "logeuclid", "riemann", "wasserstein",
+        "chol", "euclid", "logchol", "logeuclid", "riemann", "thompson",
+        "wasserstein",
         or a callable function.
 
     Returns
@@ -303,6 +386,7 @@ def geodesic(A, B, alpha, metric="riemann"):
     geodesic_logchol
     geodesic_logeuclid
     geodesic_riemann
+    geodesic_thompson
     geodesic_wasserstein
     """
     geodesic_function = check_function(metric, geodesic_functions)